Added new alternative retroarch driver.
The new driver directly waits for responses from retroarch for any given command without the need to store them in a dictionary for proper request->response coordination.

The UdpClient-wrapper also had to be reworked for this.

Pros: Should be threadsafe-ish. And there should be less work.
Cons: There's probably a higher latency for reading data.