Frontend: Avoid re-rendering of propertytree leaves.
Two parts to this: First, the PropertyService now evaluates whether the target property actually changed on call to UpdateProperty(). The Properties razor page only forwards the event if that evaluation is true.

Second: The PropertyValueViewer and PropertyTextViewer now check wether they themselves are the target of the forwarded update event.

This drastically reduces CPU load when the property editor is open. Because of that, I adjusted the batching period in the MapperClientService from 100ms to 15ms.